updated chargen menu w/ pronouns
