+++ server/conf/settings.py
@@ -35,7 +35,7 @@ from evennia.settings_default import *
 SERVERNAME = "perennia"
 AUTO_CREATE_CHARACTER_WITH_ACCOUNT = False
 AUTO_PUPPET_ON_LOGIN = False
-CHARGEN_MENU = "world.example_menu"
+CHARGEN_MENU = "world.chargen_menu"
 
 
 ######################################################################
